\name{predictBC}
\alias{predictBC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Makes predictions for a binary classification Random Forest model by averaging "probabilities". %RMR: Anna, correct?
}
\description{
This method makes predictions for a binary classification Random Forest model by computing the arithmetic mean of the "probability" 
generated by each tree, across all trees in the forest, that the instance being predicted will belong to the "selected" class. 
For a single tree, the probability is calculated as the proportion of local training set instances assigned to the terminal node 
in question which belong to the "selected" class. The class of the first instance in the complete training dataset is chosen as 
the "selected" class.  This function will only work when applied to a \code{randomForest} object  modified by 
\code{\link{prepareForPredictBC}}.   
}%RMR: Anna, is this correct?
\usage{
predictBC(object, dataT)
}

\arguments{
\item{object}{an object of class \code{randomForest}}
\item{dataT}{a data frame containing the variables in the model for the instances for which predictions are desired}
}

\value{
 A vector of predictions  for instances from the  \code{dataT} dataset. The predicted values represent the estimated probability that the instance is in the "selected" class (the class of the the first instance  in \code{dataT}).
}

\author{
Anna Palczewska \email{annawojak@gmail.com}
}

\seealso{
\code{\link[randomForest]{randomForest}}, \code{\link{prepareForPredictBC}}

}
\examples{ 
\dontrun{
library(randomForest)
data(ames)
ames_train<-ames[ames$Type=="Train",-c(1,3, ncol(ames))]
rF_Model <- randomForest(x=ames_train[,-1],y=as.factor(as.character(ames_train[,1])),
   ntree=500,importance=TRUE, keep.inbag=TRUE,replace=FALSE) 
                           
new_Model<-prepareForPredictBC(rF_Model, ames_train[,-1])
predicted<-predictBC(new_Model, ames_train[,-1])
}
}

\keyword{ feature }
\keyword{ contribution }% 
